+++ src/main/java/app/ridematrix/controller/VisitorController.java
@@ -1,18 +1,17 @@
 package app.ridematrix.controller;
 
 import app.ridematrix.dto.VisitorRequestDto;
+import app.ridematrix.dto.VisitorResponseDTO;
 import app.ridematrix.entity.Visitors;
 import app.ridematrix.service.VisitorService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Pattern;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/visitors")
@@ -22,11 +21,38 @@ public class VisitorController
     //inject service into this
     @Autowired
     private VisitorService visitorService;
+
     //API To post visitor
     @PostMapping("/addVisitor")
     @Operation(summary = "Add visitor ")
     public ResponseEntity<String> addVisitor(@Valid @RequestBody VisitorRequestDto visitorRequestDto ){
         String msg = visitorService.saveVisitor(visitorRequestDto);
         return new ResponseEntity<>(msg, HttpStatus.CREATED);
+    }
+
+    //Get visitors details with resident details by registeration number.
+    @GetMapping("/getVisitorandResidentDetails/{vehicleRegNum}")
+    @Operation(summary = "Get resident and visitor details by Vehicle Number")
+    public ResponseEntity<VisitorResponseDTO> getVisitorAndResidentDetByVehNum(
+            @Pattern(
+                    regexp = "^[A-Z0-9]{4}-[A-Z]{2}-\\d{4}$",
+                    message = "Registration number must be in format XX-00-XX-0000 equal to 10"
+            )
+            @Valid @PathVariable String vehicleRegNum) {
+        VisitorResponseDTO responseDTO = visitorService.findVisitorResidentByRegNum(vehicleRegNum);
+        return new ResponseEntity<>(responseDTO,HttpStatus.OK);
+    }
+
+    //API to update the end time of visitor by vehicle number
+    @PatchMapping("/updateOutTime/{vehicleRegNum}")
+    public ResponseEntity<String> updateTimeOut(@PathVariable String vehicleRegNum) {
+
+        String result = visitorService.updateOutTime(vehicleRegNum);
+
+        if (result != null) {
+            return ResponseEntity.ok(result);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Visitor not found with this vehicle number");
+        }
     }
 }