+++ ChangeLog.md
@@ -1,7 +1,9 @@
 # Changelog
 All notable changes to this project are documented in this file.
 ## [0.1.0] - 2025-08-1
-
+### Added
+- API- POST - new visitor details .[VEHMS-M02-T0017]
+- 
 ## [0.0.1] - 2025-08-30
 
 ### Added